// src/data/types.d.ts
export type Coin = {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  price_change_percentage_24h: number;
  total_volume: number;
};

export type PartialCoin = Partial<Coin>;

export type CoinsResponse = Coin[];